refactor(media): extract PlexSettingsPage hooks and page sections [tb-366]
Extract 2 hooks (usePlexSettings, usePlexMutations) and 5 page section
components (PlexAuthFlow, PlexMediaSync, PlexWatchSync, PlexScheduler,
PlexSyncHistory) from the 1105-line PlexSettingsPage. The main component
is now a ~134 LOC composition shell. Result display components remain
in-place for tb-367.

Closes #1713
